+++ tests/SqlArtisan.Tests/FunctionTest/FunctionTestE.cs
@@ -33,28 +33,28 @@ public partial class FunctionTest
             )
             .Build();
 
-        StringBuilder expected = new();
-        expected.Append("SELECT ");
-        expected.Append("EXTRACT(MICROSECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(SECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MINUTE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(HOUR FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DAY FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(WEEK FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MONTH FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(QUARTER FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(YEAR FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(SECOND_MICROSECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MINUTE_MICROSECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MINUTE_SECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(HOUR_MICROSECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(HOUR_SECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(HOUR_MINUTE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DAY_MICROSECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DAY_SECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DAY_MINUTE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DAY_HOUR FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(YEAR_MONTH FROM CURRENT_TIMESTAMP)");
+        StringBuilder expected = new StringBuilder()
+            .Append("SELECT ")
+            .Append("EXTRACT(MICROSECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(SECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MINUTE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(HOUR FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DAY FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(WEEK FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MONTH FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(QUARTER FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(YEAR FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(SECOND_MICROSECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MINUTE_MICROSECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MINUTE_SECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(HOUR_MICROSECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(HOUR_SECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(HOUR_MINUTE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DAY_MICROSECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DAY_SECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DAY_MINUTE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DAY_HOUR FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(YEAR_MONTH FROM CURRENT_TIMESTAMP)");
 
         Assert.Equal(expected.ToString(), sql.Text);
     }
@@ -77,19 +77,19 @@ public partial class FunctionTest
             .From(Dual)
             .Build();
 
-        StringBuilder expected = new();
-        expected.Append("SELECT ");
-        expected.Append("EXTRACT(YEAR FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(MONTH FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(DAY FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(HOUR FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(MINUTE FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(SECOND FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE_HOUR FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE_MINUTE FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE_REGION FROM SYSTIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE_ABBR FROM SYSTIMESTAMP) ");
-        expected.Append("FROM DUAL");
+        StringBuilder expected = new StringBuilder()
+            .Append("SELECT ")
+            .Append("EXTRACT(YEAR FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(MONTH FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(DAY FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(HOUR FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(MINUTE FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(SECOND FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE_HOUR FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE_MINUTE FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE_REGION FROM SYSTIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE_ABBR FROM SYSTIMESTAMP) ")
+            .Append("FROM DUAL");
 
         Assert.Equal(expected.ToString(), sql.Text);
     }
@@ -108,6 +108,7 @@ public partial class FunctionTest
                 Extract(DatePart.Hour, CurrentTimestamp),
                 Extract(DatePart.Isodow, CurrentTimestamp),
                 Extract(DatePart.Isoyear, CurrentTimestamp),
+                Extract(DatePart.Julian, CurrentTimestamp),
                 Extract(DatePart.Microseconds, CurrentTimestamp),
                 Extract(DatePart.Millennium, CurrentTimestamp),
                 Extract(DatePart.Milliseconds, CurrentTimestamp),
@@ -123,29 +124,30 @@ public partial class FunctionTest
             )
             .Build();
 
-        StringBuilder expected = new();
-        expected.Append("SELECT ");
-        expected.Append("EXTRACT(CENTURY FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DAY FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DECADE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DOW FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(DOY FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(EPOCH FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(HOUR FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(ISODOW FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(ISOYEAR FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MICROSECONDS FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MILLENNIUM FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MILLISECONDS FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MINUTE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(MONTH FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(QUARTER FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(SECOND FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE_HOUR FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(TIMEZONE_MINUTE FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(WEEK FROM CURRENT_TIMESTAMP), ");
-        expected.Append("EXTRACT(YEAR FROM CURRENT_TIMESTAMP)");
+        StringBuilder expected = new StringBuilder()
+            .Append("SELECT ")
+            .Append("EXTRACT(CENTURY FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DAY FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DECADE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DOW FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(DOY FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(EPOCH FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(HOUR FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(ISODOW FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(ISOYEAR FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(JULIAN FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MICROSECONDS FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MILLENNIUM FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MILLISECONDS FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MINUTE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(MONTH FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(QUARTER FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(SECOND FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE_HOUR FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(TIMEZONE_MINUTE FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(WEEK FROM CURRENT_TIMESTAMP), ")
+            .Append("EXTRACT(YEAR FROM CURRENT_TIMESTAMP)");
 
         Assert.Equal(expected.ToString(), sql.Text);
     }