+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.1.0-alpha.17] - 2025-06-16
 ### Added
 - Added support for `EXTRACT` function.
+- Added support for the SQL Server `DATEPART` function.
 - Added a "Why SqlArtisan?" section to the README.
 
 ## [0.1.0-alpha.16] - 2025-06-11